Dependency Injection, Small Refactoring
- Add Dependency Injection to Event Receiver project;
- Clean up code;

+++ MarineTask/MarineTask.EventSender/PublishDetails.cs
@@ -6,7 +6,7 @@
 
         public int BlockNumber { get; set; }
 
-        public long BlockCount { get; set; }
+        public int BlockCount { get; set; }
 
         public byte[] Data { get; set; }
     }
